# Conflicts:
#	package.json
#	src/core/router/index.tsx
#	src/pages/firstrun/FirstRunPage.tsx
#	src/pages/firstrun/Footer.tsx
#	src/pages/main/MainPage.tsx
#	yarn.lock

+++ run.js
@@ -95,7 +95,8 @@ tasks.set('bundle', () => {
 tasks.set('release', () => {
   const zipFolder = require('zip-a-folder');
 
-  fs.rmdirSync('./public/dist/sourcemaps', { recursive: true });
+  if (fs.existsSync('./public/dist/sourcemaps'))
+    fs.rmdirSync('./public/dist/sourcemaps', { recursive: true });
   return zipFolder.zip('./public', process.env.NODE_ENV === 'development' ? './build/latest-unstable.zip' : './build/latest.zip')
     .then(() => { console.log('Release build created!'); });
 });
